+++ src/plugins/index.ts
@@ -2,4 +2,5 @@ export * from './dataLoader';
 export * from './gridInitializer';
 export * from './metadataSelector';
 export * from './noDataRenderer';
+export * from './paging';
 export * from './rowSelector';