+++ src/plugins/paging/basic/basicPaging.interface.ts
@@ -1,14 +1,14 @@
-import {PagingOptions, Paging} from '../paging.interface';
+import {Paging, PagingOptions} from '../../../interfaces';
 
 /**
  * Css classes for basic paging
  */
 export interface CssClassesBasicPaging
 {
-    pagingUl?: string;
-    itemsPerPageDiv?: string;
-    displayedItemsCountSpan?: string;
-    itemsPerPageUl?: string;
+    pagingUl: string;
+    itemsPerPageDiv: string;
+    displayedItemsCountSpan: string;
+    itemsPerPageUl: string;
 }
 
 /**
@@ -19,12 +19,12 @@ export interface BasicPagingOptions extends PagingOptions<CssClassesBasicPaging>
     /**
      * Page dispersion parameter for rendered pages
      */
-    pagesDispersion?: number;
+    pagesDispersion: number;
 
     /**
      * Gets or sets array of available values for itemsPerPage, NaN represents infinity (all data)
      */
-    itemsPerPageValues?: number[];
+    itemsPerPageValues: number[];
 }
 
 /**