+++ src/plugins/paging/index.ts
@@ -1,0 +1,5 @@
+export * from './basic/basicPaging.component';
+export * from './basic/basicPaging.interface';
+export * from './no/noPaging.component';
+export * from './no/noPaging.interface';
+export * from './pagingAbstract.component';