+++ src/plugins/paging/previousNext/previousNextPaging.component.ts
@@ -110,7 +110,7 @@ export class PreviousNextPagingComponent extends PagingAbstractComponent<CssClas
      */
     public get initialized(): boolean
     {
-        return this._initialized;
+        return this.initialized;
     }
 
     //######################### public properties - inputs #########################
@@ -166,7 +166,7 @@ export class PreviousNextPagingComponent extends PagingAbstractComponent<CssClas
     {
         super(pluginElement, changeDetector, gridPlugins);
 
-        this._options = extend(true, {}, defaultOptions, options);
+        this.ɵoptions = extend(true, {}, defaultOptions, options);
         this.optionsSet();
     }