+++ src/plugins/paging/virtualScroll/virtualScrollPagingAbstract.component.ts
@@ -130,7 +130,7 @@ export abstract class VirtualScrollPagingAbstractComponent<TOptions extends Virt
 
         const offset = element.scrollTop / (element.scrollHeight - element.clientHeight);
 
-        if(isNaN(offset) || offset >= this._options.loadOffsetTreshold)
+        if(isNaN(offset) || offset >= this.ɵoptions.loadOffsetTreshold)
         {
             this._load();
         }